add some jank code to draw every unit.

+++ core/src/main/java/com/ziodyne/sometrpg/view/entities/EntityFactory.java
@@ -73,7 +73,7 @@ public class EntityFactory {
     int frameDims = 40;
     Array<TextureRegion> regions = new Array<>();
     for (Vector2 frameCoord : spec.getFrameCoords()) {
-      regions.add(new TextureRegion(texture, (int)frameCoord.x*frameDims, (int)frameCoord.y*frameDims, frameDims, frameDims));
+      regions.add(new TextureRegion(texture, (int)frameCoord.x*frameDims, frameCoord.y*frameDims, frameDims, frameDims));
     }
 
     Animation animation = new Animation(spec.getFrameDurationMs() / 1000f, regions, Animation.LOOP);
